+++ src/main/java/org/organicdesign/fp/collections/PersistentTreeSet.java
@@ -29,7 +29,7 @@ import static org.organicdesign.fp.collections.Equator.defaultComparator;
  This file is a derivative work based on a Clojure collection licensed under the Eclipse Public
  License 1.0 Copyright Rich Hickey.  Errors by Glen Peterson.
  */
-public class PersistentTreeSet<E> extends UnmodSortedSet.AbstractUnmodSet<E>
+public class PersistentTreeSet<E> extends AbstractUnmodSet<E>
         implements ImSortedSet<E>, Serializable {
 
     /**
@@ -231,7 +231,7 @@ public class PersistentTreeSet<E> extends UnmodSortedSet.AbstractUnmodSet<E>
 
     /** {@inheritDoc} */
     @Override public Option<E> head() {
-        return size() > 0 ? Option.of(impl.firstKey()) : Option.none();
+        return size() > 0 ? Option.some(impl.firstKey()) : Option.none();
     }
 
     /** {@inheritDoc} */