+++ src/main/java/org/organicdesign/fp/collections/UnmodIterable.java
@@ -15,17 +15,6 @@ import static org.organicdesign.fp.FunctionUtils.stringify;
 public interface UnmodIterable<T> extends Iterable<T>, Transformable<T> {
     // ========================================== Static ==========================================
 
-    /**
-     Implements equals and hashCode() methods compatible with all java.util collections (this
-     algorithm is not order-dependent) and toString which takes the name of the sub-class.
-     */
-    abstract class AbstractUnmodIterable<T> implements UnmodIterable<T> {
-        @Override public int hashCode() { return UnmodIterable.hash(this); }
-        @Override public String toString() {
-            return UnmodIterable.toString(getClass().getSimpleName(), this);
-        }
-    }
-
     //    /**
 // Caution: this is a convenient optimization for immutable data structures and a nightmare
 // waiting to happen to mutable ones.  Don't go slapping this on immutable wrappers for mutable
@@ -129,9 +118,9 @@ public interface UnmodIterable<T> extends Iterable<T>, Transformable<T> {
             sB.append(stringify(item));
             i++;
         }
-        if (iter.hasNext()) {
-            sB.append("...");
-        }
+//        if (iter.hasNext()) {
+//            sB.append("...");
+//        }
         return sB.append(")").toString();
     }
 
@@ -206,7 +195,7 @@ public interface UnmodIterable<T> extends Iterable<T>, Transformable<T> {
      */
     default Option<T> head() {
         Iterator<T> iter = iterator();
-        return iter.hasNext() ? Option.of(iter.next())
+        return iter.hasNext() ? Option.some(iter.next())
                               : Option.none();
     }