+++ CHANGE_LOG.md
@@ -18,7 +18,8 @@ Major changes:
  - Added JavaDoc to Git site
  - Changed toString to show entire collection
  - Added rrb() and mutableRrb() to StaticImports for constructing RRB Trees.
- - Moved AbstractUnmodMap from inner class of UnmodMap to top-level class (ebean complained about the inner class).
+ - Moved AbstractUnmodMap, AbstractUnmodSet, and AbstractUnmodIterable from inner classes to top-level classes.
+ - Option.of() is now Option.some() and it no longer returns None if you pass it a None.
 
 Here is a script to ease your upgrade from 2.1.1 to 3.0:
 ```bash